chore: code separation
