+++ extensions/cursor-recent-projects/src/integrations/types.ts
@@ -1,0 +1,11 @@
+import { LaunchOptions } from "raycast-cross-extension";
+
+export interface CursorDirectoryContext {
+  ruleContent: string;
+  replace?: boolean;
+}
+
+export type LaunchContext = {
+  cursorDirectory?: Partial<CursorDirectoryContext>;
+  callbackLaunchOptions?: LaunchOptions;
+};
